Redesign tealdgb
* There are three main components:
1. Debugger
2. Hook adapters - event receivers from TEAL eval
3. Debug adapters - frontend protocol support (web, CDT, VSCode)

* Currently supported CDT and webpage
