+++ cmd/tealdbg/util.go
@@ -1,0 +1,75 @@
+// Copyright (C) 2019-2020 Algorand, Inc.
+// This file is part of go-algorand
+//
+// go-algorand is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// go-algorand is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with go-algorand.  If not, see <https://www.gnu.org/licenses/>.
+
+package main
+
+import (
+	"sync/atomic"
+)
+
+type atomicString struct {
+	value atomic.Value
+}
+
+func (s *atomicString) Store(other string) {
+	s.value.Store(other)
+}
+
+func (s *atomicString) Load() string {
+	result := s.value.Load()
+	if result != nil {
+		if value, ok := result.(string); ok {
+			return value
+		}
+	}
+	return ""
+}
+
+func (s *atomicString) Length() int {
+	return len(s.Load())
+}
+
+type atomicBool struct {
+	value uint32
+}
+
+func (b *atomicBool) SetTo(other bool) {
+	var converted uint32 = 0
+	if other {
+		converted = 1
+	}
+	atomic.StoreUint32(&b.value, converted)
+}
+
+func (b *atomicBool) IsSet() bool {
+	return atomic.LoadUint32(&b.value) != 0
+}
+
+type atomicInt struct {
+	value int32
+}
+
+func (i *atomicInt) Store(other int) {
+	atomic.StoreInt32(&i.value, int32(other))
+}
+
+func (i *atomicInt) Load() int {
+	return int(atomic.LoadInt32(&i.value))
+}
+
+func (i *atomicInt) Add(other int) int {
+	return int(atomic.AddInt32(&i.value, int32(other)))
+}
